+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Update code style
 * Update composer requirements
 * Fix broken link in README
+* Add filter `lcfirst`
 
 ## 1.0.3
 * Update minimal PHP requirement to 5.3.3